+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 5.0.0-alpha.0
 * Use CSS instead of mui's styles mechanism.
 * Use material-ui 5.x alpha.
 * Load react-markdown once a description dialog is first opened.